% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_line_chart.R
\name{create_line_chart}
\alias{create_line_chart}
\title{create_line_chart}
\usage{
create_line_chart(data, id_selected, metric_1, metric_2)
}
\arguments{
\item{data}{A dataframe to create the line chart from}

\item{id_selected}{Participant ID selected to filter the line chart by}

\item{metric_1}{The first metric to show in the line chart}

\item{metric_2}{The second metric to show in the line chart}
}
\value{
a ggplot line chart
}
\description{
create a rolling mean line chart
}
