% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_sankey.R
\name{create_participant_sankey}
\alias{create_participant_sankey}
\title{create_participant_sankey}
\usage{
create_participant_sankey(data, participant)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}

\item{participant}{The participant ID of the participant to create the sankey diagram for}
}
\value{
A highcharter sankey diagram
}
\description{
A function to create a highcharter sankey diagram for individual participants
}
