% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{create_sankey}
\alias{create_sankey}
\title{create_sankey}
\usage{
create_sankey(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A highcharter Sankey diagram
}
\description{
A function to create a sankey diagram
}
