% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_snack_count.R
\name{daily_snack_count}
\alias{daily_snack_count}
\title{daily_snack_count}
\usage{
daily_snack_count(data)
}
\arguments{
\item{data}{A dataframe to provide to the function}
}
\value{
A dataframe summarising SnackApp screen visits daily
}
\description{
A function to count the number of times a participant visits each app screen in the SnackApp
}
