% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_qaqc.R
\name{dataload_qaqc}
\alias{dataload_qaqc}
\title{dataload_qaqc}
\usage{
dataload_qaqc(path, name)
}
\arguments{
\item{path}{Path to folder that contains the data files}

\item{name}{}
}
\value{
A QAQC dataframe summarising the SnackApp usage data uploaded
}
\description{

}
