% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_qaqc_fb_usage.R
\name{dataload_qaqc_fb_usage}
\alias{dataload_qaqc_fb_usage}
\title{dataload_qaqc_fb_usage}
\usage{
dataload_qaqc_fb_usage(path, name)
}
\arguments{
\item{path}{path to folder where data files are stored}

\item{name}{}
}
\value{
A QAQC dataframe summarising the Fitbit usage data files uploaded
}
\description{
dataload_qaqc_fb_usage
}
