% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_pa_qaqc.R
\name{fb_pa_dataload_qaqc}
\alias{fb_pa_dataload_qaqc}
\title{fb_pa_dataload_qaqc}
\usage{
fb_pa_dataload_qaqc(path, name)
}
\arguments{
\item{path}{Path to the folder containing the Fitbit physical actvity data files}

\item{name}{}
}
\value{
A QAQC dataframe summarising the files uploaded
}
\description{
A function to run QAQC on the Fitbit physical activity data uploaded
}
