% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_usage_day_summary.R
\name{fb_usage_day_summary}
\alias{fb_usage_day_summary}
\title{fb_usage_day_summary}
\usage{
fb_usage_day_summary(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe summarising participants daily Fitbit usage
}
\description{
A function to produce a daily summary of Fitbit usage data
}
