% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_usage_qaqc.R
\name{fb_usage_full_qaqc}
\alias{fb_usage_full_qaqc}
\title{fb_usage_full_qaqc}
\usage{
fb_usage_full_qaqc(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe containing full QAQC data on Fitbit usage
}
\description{
A function to run a complete QAQC for Fitbit usage
}
