% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_usage_participant_summary.R
\name{fb_usage_participant_summary}
\alias{fb_usage_participant_summary}
\title{fb_usage_participant_summary}
\usage{
fb_usage_participant_summary(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe summarising each participants Fitbit usage (one row per participant)
}
\description{
A function to produce a summary per participant of Fitbit usage
}
