% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_pa_daily.R
\name{fitbit_pa_daily}
\alias{fitbit_pa_daily}
\title{fitbit_pa_daily}
\usage{
fitbit_pa_daily(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe summarising Fitbit physicalactivity metrics daily
}
\description{
A function to summarise Fitbit physical activity daily
}
