% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_pa_load_data.R
\name{fitbit_pa_load_data}
\alias{fitbit_pa_load_data}
\title{fitbit_pa_load_data}
\usage{
fitbit_pa_load_data(path, name)
}
\arguments{
\item{path}{The path to the folder containing the Fitbit physical activity data files}

\item{name}{}
}
\value{
A dataframe of Fitbit physical activity data
}
\description{
A function to load Fitbit physical activity data
}
