% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_pa_month.R
\name{fitbit_pa_month}
\alias{fitbit_pa_month}
\title{fitbit_pa_month}
\usage{
fitbit_pa_month(data)
}
\arguments{
\item{data}{A dataframe of Fitbit physical activity data}
}
\value{
A dataframe summarising Fitbit physical activity data monthly
}
\description{
A function to summarise Fitbit physical activity data at a monthly level
}
