% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_pa_qaqc.R
\name{fitbit_pa_qaqc}
\alias{fitbit_pa_qaqc}
\title{fitbit_pa_qaqc}
\usage{
fitbit_pa_qaqc(data)
}
\arguments{
\item{data}{A dataframe containing Fitbit physical activity data}
}
\value{
A dataframe containing QAQC information
}
\description{
A function to perform QAQC on Fitbit physical activity data
}
