% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_pa_week.R
\name{fitbit_pa_week}
\alias{fitbit_pa_week}
\title{fitbit_pa_week}
\usage{
fitbit_pa_week(data)
}
\arguments{
\item{data}{A dataframe of Fitbit physical activity data}
}
\value{
Fitbit physical activity data summarised weekly
}
\description{
A function to summarise Fitbit physical activity data weekly
}
