% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(path, name, id_start, id_end)
}
\arguments{
\item{path}{The path to the folder containing SnackApp usage data files}

\item{name}{}

\item{id_start}{The numeric location where the participant ID starts in each data file}

\item{id_end}{The numeric location where the participant ID ends in each data file}
}
\value{
A dataframe containing SnackApp usage data
}
\description{
A function to load SnackApp usage data
}
