% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fb_data.R
\name{load_fb_data}
\alias{load_fb_data}
\title{load_fb_data}
\usage{
load_fb_data(path, name)
}
\arguments{
\item{path}{The path to the folder containing the Fitbit usage data files}

\item{name}{}
}
\value{
A dataframe containing Fitbit usage data
}
\description{
A function to load Fitbit usage data
}
