% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_snack_count.R
\name{participant_snack_count}
\alias{participant_snack_count}
\title{Title participant_snack_count}
\usage{
participant_snack_count(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A dataframe of summary information regarding SnackApp usage
}
\description{
A function to count the number of screens viewed by a participant within the SnackApp
}
