% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qaqc}
\alias{qaqc}
\title{qaqc}
\usage{
qaqc(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A QAQC summary of SnackApp usage data
}
\description{
a function to perform QAQC on the SnackApp usage summary
}
