% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackapp_usage_week_summary.R
\name{seven_day_summary}
\alias{seven_day_summary}
\title{seven_day_summary}
\usage{
seven_day_summary(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A dataframe summarising SnackApp usage data weekly
}
\description{
A function to generate a summary of SnackApp usage data weekly
}
