% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackapp_usage_daily_sum.R
\name{snackapp_usage_daily_sum}
\alias{snackapp_usage_daily_sum}
\title{snackapp_usage_daily_sum}
\usage{
snackapp_usage_daily_sum(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A summary dataframe of SnackApp usage data
}
\description{
A function to create a summary of SnackApp usage data daily
}
