% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackapp_usage_p_sum.R
\name{snackapp_usage_p_sum}
\alias{snackapp_usage_p_sum}
\title{snackapp_usage_p_sum}
\usage{
snackapp_usage_p_sum(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A dataframe summarising SnackApp usage data per participant (one row per participant)
}
\description{
A function to generate a summary of SnackApp usage data for each participant
}
