% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_snack_count.R
\name{weekly_snack_count}
\alias{weekly_snack_count}
\title{weekly_snack_count}
\usage{
weekly_snack_count(data)
}
\arguments{
\item{data}{A dataframe of SnackApp usage data}
}
\value{
A dataframe summarising SnackApp usage weekly
}
\description{
A function to generate a weekly summary of SnackApp usage data
}
